# trader_interactive/public/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "L5 Swagger API",
        "description": "L5 Swagger API description",
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/sellers": {
            "get": {
                "tags": [
                    "sellers"
                ],
                "summary": "Get list of sellers",
                "description": "Returns list of sellers",
                "operationId": "getSellers",
                "responses": {
                    "200": {
                        "description": "Sellers Details Found"
                    },
                    "500": {
                        "description": "Sellers Details Not Found"
                    }
                }
            }
        },
        "/api/seller-details/id={id}": {
            "get": {
                "tags": [
                    "seller-details"
                ],
                "summary": "Get Individual Seller Details",
                "description": "Returns Individual Seller Details",
                "operationId": "getSellerDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "seller id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Individual Seller Details Found"
                    },
                    "500": {
                        "description": "Individual Seller Details Not Found"
                    }
                }
            }
        },
        "/api/seller-listing-details/id={id}": {
            "get": {
                "tags": [
                    "seller-listing-details"
                ],
                "summary": "Get Seller Listing Details",
                "description": "Returns Seller Listing Details",
                "operationId": "getIndividualSellerDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "seller id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Seller Listing Details Found"
                    },
                    "500": {
                        "description": "Seller Listing Details Not Found"
                    }
                }
            }
        },
        "/api/seller-individual-listing-details/id={id}/type={type}": {
            "get": {
                "tags": [
                    "seller-individual-listing-details"
                ],
                "summary": "Get Seller Individual Listing Details",
                "description": "Returns Seller Individual Listing Details",
                "operationId": "getSellerListingDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "seller id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "type",
                        "in": "path",
                        "description": "listing type",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Seller Individual Listing Details Found"
                    },
                    "500": {
                        "description": "Seller Individual Listing Details Not Found"
                    }
                }
            }
        },
        "/api/create-sellers": {
            "post": {
                "tags": [
                    "create-sellers"
                ],
                "summary": "Create sellers",
                "description": "Create sellers",
                "operationId": "createSellers",
                "parameters": [
                    {
                        "name": "createSellers",
                        "in": "body",
                        "description": "createSellers",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "example": {
                                "type": "Dealer",
                                "name": "DealerTest",
                                "address": "swagger",
                                "phone": "987654321",
                                "email": "john@example.com",
                                "website": "ohn@example.com"
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sellers Created Successfully"
                    },
                    "500": {
                        "description": "Sellers Not Created"
                    }
                }
            }
        },
        "/api/create-sellers-listings": {
            "post": {
                "tags": [
                    "create-sellers-listings"
                ],
                "summary": "Create sellers listings",
                "description": "Create sellers listings",
                "operationId": "createSellersListings",
                "parameters": [
                    {
                        "name": "seller_listings",
                        "in": "body",
                        "description": "seller_listings",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "example": {
                                "seller_id": "1",
                                "listing_id": "2",
                                "selling_price": "2131432"
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Seller's Listings Created Successfully"
                    },
                    "500": {
                        "description": "Seller's Listings Not Created"
                    }
                }
            }
        },
...
    },
...
}